Set some logics update the splash screen review the code
