+++ SarmadAndroidAppCollection/Blog App/app/src/main/java/com/sarmadtechempire/blogapp/MainActivity.java
@@ -21,7 +21,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-
         // Initialize view binding for the main activity layout
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
