+++ SarmadAndroidAppCollection/Blog App/app/src/main/java/com/sarmadtechempire/blogapp/SplashActivity.java
@@ -12,11 +12,11 @@ import androidx.core.view.WindowInsetsCompat;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.sarmadtechempire.blogapp.register.RegisterActivity;
-import com.sarmadtechempire.blogapp.register.LoginActivity;
 import com.sarmadtechempire.blogapp.register.WelcomeActivity;
 
 public class SplashActivity extends AppCompatActivity {
+
+    private FirebaseAuth auth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -24,16 +24,7 @@ public class SplashActivity extends AppCompatActivity {
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_splash);
 
-
-        Intent iHome = new Intent(SplashActivity.this, WelcomeActivity.class);
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                startActivity(iHome);
-                finish();
-
-            }
-        },4000);
+        auth = FirebaseAuth.getInstance();
 
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
@@ -42,4 +33,34 @@ public class SplashActivity extends AppCompatActivity {
         });
     }
 
+    @Override
+    protected void onStart() {
+        super.onStart();
+        // Note: Since there is no progress bar in the splash screen layout, we are not showing/hiding it.
+
+        FirebaseUser currentUser = auth.getCurrentUser();
+
+        if (currentUser != null) {
+            navigateToMainScreen();
+        } else {
+            navigateToWelcomeScreen();
+        }
+    }
+
+    private void navigateToMainScreen() {
+        Intent intent = new Intent(SplashActivity.this, MainActivity.class);
+        startActivity(intent);
+        finish();
+    }
+
+    private void navigateToWelcomeScreen() {
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent iHome = new Intent(SplashActivity.this, WelcomeActivity.class);
+                startActivity(iHome);
+                finish();
+            }
+        }, 4000); // Splash screen duration
+    }
 }
